+++ src/debvader/deblend_cutout/deblender.py
@@ -1,0 +1,25 @@
+import numpy as np
+import pandas as pd
+import scipy
+import tensorflow as tf
+
+
+
+def deblend(net, images, normalised=False):
+    """
+    Deblend the image using the network
+    parameters:
+        net: neural network used to do the deblending
+        images: array of images. It can contain only one image
+        normalised: boolean to indicate if images need to be normalised
+    """
+    if normalised:
+        # Normalize input images
+        images_normed = np.tanh(np.arcsinh(images))
+        # Denorm output images
+        images = np.sinh(np.arctanh(net.predict(images_normed)))
+
+    return net(tf.cast(images, tf.float32)).mean().numpy(), net(
+        tf.cast(images, tf.float32)
+    )
+
