+++ src/debvader/deblend_cutout/model.py
@@ -18,7 +18,7 @@ from tensorflow.keras.layers import (
 )
 from tensorflow.keras.models import Model
 
-from debvader.deblend.metrics import vae_loss
+from debvader.deblend_cutout.metrics import vae_loss
 
 tfd = tfp.distributions
 tfb = tfp.bijectors
